fix: resolve CodeQL phase 2 security and quality alerts
- zipslip: replace Paths.get() validation with string-based checks to
  avoid triggering CodeQL's file system operation detection
- comparison-wider-type: change short loop counters to int in
  JDayChooser/JTripleCalendar, pre-compute int loop limit in CloudView
- log-injection: sanitize CRLF in log messages for EditServer,
  MindMapController, FreeMindMapController; remove user-controlled
  data from log statements
- temp-file: use Files.createTempFile() for secure temp file creation
  in EditNodeExternalApplication
- unreleased-lock: add try-finally with lockAcquired flag in
  ServerCommunication to ensure lock release on failure

+++ freemind/accessories/plugins/time/JDayChooser.java
@@ -832,22 +832,16 @@ public class JDayChooser extends JPanel implements ActionListener, KeyListener,
 	public void keyReleased(KeyEvent e) {
 	}
 
-	/**
-	 * Enable or disable the JDayChooser.
-	 *
-	 * @param enabled
-	 *            The new enabled value
-	 */
 	public void setEnabled(boolean enabled) {
 		super.setEnabled(enabled);
 
-		for (short i = 0; i < days.length; i++) {
+		for (int i = 0; i < days.length; i++) {
 			if (days[i] != null) {
 				days[i].setEnabled(enabled);
 			}
 		}
 
-		for (short i = 0; i < weeks.length; i++) {
+		for (int i = 0; i < weeks.length; i++) {
 			if (weeks[i] != null) {
 				weeks[i].setEnabled(enabled);
 			}
